trialmatch/backend/requirements.txt
# TrialMatch Backend — Dépendances Python 3.11

# ── Framework ──────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Base de données ────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ── Validation ────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Sécurité ──────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7
bcrypt==4.1.3

# ── HTTP Client (Ollama) ──────────────────────────────────────────────────────
httpx==0.27.0

# ── LDAP (optionnel) ──────────────────────────────────────────────────────────
python-ldap==3.4.4; sys_platform != "win32"

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Logging structuré ─────────────────────────────────────────────────────────
structlog==24.1.0

# ── Tests ─────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0
faker==25.0.0
respx==0.21.1
